+++ plone/app/contenttypes/utils.py
@@ -42,18 +42,3 @@ def replace_link_variables_by_paths(context, url):
 def _replace_variable_by_path(url, variable, obj):
     path = '/'.join(obj.getPhysicalPath())
     return url.replace(variable, path)
-
-
-def human_readable_size(size):
-    """Return human readable size.
-    Based on https://stackoverflow.com/a/1094933
-    """
-    if size < 1024:
-        return '{size} B'.format(size=size)
-
-    for unit in ('KB', 'MB', 'GB'):
-        size /= 1024.0
-        if size < 1024.0:
-            return '{size:.1f} {unit}'.format(size=size, unit=unit)
-
-    return '{size:.1f} GB'.format(size=size)